Move some dust functions to a separate module and clean up imports in data_vectors.py
